Update matrix multiplication

+++ tutQ3.c
@@ -63,7 +63,7 @@ int* rank2TensorMult(int *matrixA, int dimsA[2], int *matrixB, int dimsB[2], int
 			for(int mult_op=0; mult_op<total_multiplication_operations; mult_op++)
 			{
 				// Get matrix A indexes:
-				int tmp = mult_op/(total_multiplication_operations/rows_in_A);
+				int tmp = mult_op/(float)(total_multiplication_operations/(float)rows_in_A);
 				if(tmp % rows_in_A) rowA = tmp;
 				colA = mult_op % cols_in_A;
